requirements.txt
streamlit
langchain-openai
langchain
python-dotenv
langchain_community
